Add interface tests
